Restructure frontend project paths
