+++ frontend/src/Home.tsx
@@ -1,6 +1,6 @@
 import { Typography } from "@material-ui/core";
 import React from "react";
-import Header from "./Header";
+import Header from "./components/Header";
 
 function Home() {
   return (
